chore: [KD-] add sample dashboard helpers for headless smoke flow
Dev URL overrides, dashboard state, and logging for SDK sample app.
